:truck: [#31] Move files around to declutter
Split up the various components/aspects of the date constraint
configuration component into separate files.
